# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "trao interview prep kit",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/kits/{kit_id}/practice/queue": {
      "get": {
        "summary": "Queue",
        "operationId": "queue_api_kits__kit_id__practice_queue_get",
        "parameters": [
          {
            "name": "kit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kit Id"
            }
          },
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/kits/{kit_id}/practice/reviews": {
      "post": {
        "summary": "Review",
        "operationId": "review_api_kits__kit_id__practice_reviews_post",
        "parameters": [
          {
            "name": "kit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kit Id"
            }
          },
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Review"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/kits/{kit_id}/practice/summary": {
      "get": {
        "summary": "Summary",
        "operationId": "summary_api_kits__kit_id__practice_summary_get",
        "parameters": [
          {
            "name": "kit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kit Id"
            }
          },
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PracticeSummaryOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/kits/{kit_id}/practice/weak-spots": {
      "get": {
        "summary": "Weak Spots",
        "description": "Rank Requirements by readiness with reasons (practice Confidence, coverage, priority).",
        "operationId": "weak_spots_api_kits__kit_id__practice_weak_spots_get",
        "parameters": [
          {
            "name": "kit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kit Id"
            }
          },
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeakSpotsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/kits/{kit_id}/practice/check": {
      "post": {
        "summary": "Check",
        "operationId": "check_api_kits__kit_id__practice_check_post",
        "parameters": [
          {
            "name": "kit_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kit Id"
            }
          },
          {
            "name": "session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CheckBody": {
        "properties": {
          "question_id": {
            "type": "string",
            "title": "Question Id"
          },
          "answer": {
            "type": "string",
            "title": "Answer"
          }
        },
        "type": "object",
        "required": [
          "question_id",
          "answer"
        ],
        "title": "CheckBody"
      },
      "CheckOut": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/CheckPoint"
            },
            "type": "array",
            "title": "Results"
          },
          "method": {
            "type": "string",
            "title": "Method"
          },
          "limits": {
            "type": "string",
            "title": "Limits"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "results",
          "method",
          "limits"
        ],
        "title": "CheckOut"
      },
      "CheckPoint": {
        "properties": {
          "point": {
            "type": "string",
            "title": "Point"
          },
          "covered": {
            "type": "boolean",
            "title": "Covered"
          }
        },
        "type": "object",
        "required": [
          "point",
          "covered"
        ],
        "title": "CheckPoint"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OkOut": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          }
        },
        "type": "object",
        "title": "OkOut"
      },
      "PracticeSummaryOut": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "covered": {
            "type": "integer",
            "title": "Covered"
          },
          "uncovered": {
            "type": "integer",
            "title": "Uncovered"
          }
        },
        "type": "object",
        "required": [
          "total",
          "covered",
          "uncovered"
        ],
        "title": "PracticeSummaryOut"
      },
...
      "QueueOut": {
        "properties": {
          "queue": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Queue"
          }
        },
        "type": "object",
        "required": [
          "queue"
        ],
        "title": "QueueOut"
      },
...
      "Review": {
        "properties": {
          "flashcard_id": {
            "type": "string",
            "title": "Flashcard Id"
          },
          "confidence": {
            "type": "integer",
            "title": "Confidence"
          }
        },
        "type": "object",
        "required": [
          "flashcard_id",
          "confidence"
        ],
        "title": "Review"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WeakSpotEntry": {
        "properties": {
          "requirement_id": {
            "type": "string",
            "title": "Requirement Id"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reasons"
          },
          "score": {
            "type": "integer",
            "title": "Score"
          },
          "question_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Question Ids"
          },
          "flashcard_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Flashcard Ids"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "requirement_id",
          "reasons",
          "score",
          "question_ids",
          "flashcard_ids"
        ],
        "title": "WeakSpotEntry"
      },
      "WeakSpotsOut": {
        "properties": {
          "spots": {
            "items": {
              "$ref": "#/components/schemas/WeakSpotEntry"
            },
            "type": "array",
            "title": "Spots"
          }
        },
        "type": "object",
        "required": [
          "spots"
        ],
        "title": "WeakSpotsOut"
      }
    }
  }
}